industrial_ai_chatbot_v1.1/requirements.txt
streamlit>=1.31.0
anthropic>=0.18.0
openai>=1.12.0
google-genai>=0.2.0
python-dotenv>=1.0.0
